jobqueue: move simple logger to pkg/
While jobqueue is exposed via pkg/ package, the logging facade was still in the internal/ package making it impossible to actually import the interface. This patch moves the simple logger facade into a newly-created package in pkg/ and also moves implementations into separate packages so the interface can be imported without loading dependencies.
