+++ pkg/jobqueue/jobqueue.go
@@ -79,18 +79,6 @@ type JobQueue interface {
 	RefreshHeartbeat(token uuid.UUID)
 }
 
-// SimpleLogger provides a structured logging methods for the jobqueue library.
-type SimpleLogger interface {
-	// Info creates an info-level message and arbitrary amount of key-value string pairs which
-	// can be optionally mapped to fields by underlying implementations.
-	Info(msg string, args ...string)
-
-	// Error creates an error-level message and arbitrary amount of key-value string pairs which
-	// can be optionally mapped to fields by underlying implementations. The first error argument
-	// can be set to nil when no context error is available.
-	Error(err error, msg string, args ...string)
-}
-
 var (
 	ErrNotExist       = errors.New("job does not exist")
 	ErrNotPending     = errors.New("job is not pending")
